+++ cmd/hla/hla-http/daemon/install_hla_http.sh
@@ -6,7 +6,7 @@ echo "
 Description=HiddenLakeAdapterHTTP
 
 [Service]
-ExecStart=/usr/local/bin/hla-http_amd64_linux --path ~/.config/hidden-lake
+ExecStart=/usr/local/bin/hla-http_amd64_linux --path /usr/local/etc/hidden-lake
 Restart=always
 RestartSec=10