+++ cmd/hlc/daemon/install_hlc.sh
@@ -6,7 +6,7 @@ echo "
 Description=HiddenLakeComposite
 
 [Service]
-ExecStart=/usr/local/bin/hlc_amd64_linux --path ~/.config/hidden-lake
+ExecStart=/usr/local/bin/hlc_amd64_linux --path /usr/local/etc/hidden-lake
 Restart=always
 RestartSec=10